refactor: extract getDayCellColorHandler from CustomerCalendar and InstructorCalendarClient
